+++ spec/selectors/region_spec.rb
@@ -3,20 +3,21 @@
 describe "region selector" do
   describe "locator" do
     context "when <section> element" do
-      it "finds the first element" do
+      it "does not find with no accessible label" do
         render <<~HTML
           <section>Content</section>
         HTML
 
-        expect(page).to have_selector :region, count: 1
+        expect(page).to have_no_selector :region
       end
 
       it "finds based on [aria-label]" do
         render <<~HTML
           <section aria-label="Section region">Content</section>
+          <section aria-label="Other region">Content</section>
         HTML
 
-        expect(page).to have_selector :region, "Section region"
+        expect(page).to have_selector :region, "Section region", count: 1
       end
 
       it "does not find differing on [aria-label]" do
@@ -32,9 +33,12 @@ describe "region selector" do
           <section aria-labelledby="section_label">
             <h1 id="section_label">Section region</h1>
           </section>
+          <section aria-labelledby="section_other">
+            <h1 id="section_other">Other region</h1>
+          </section>
         HTML
 
-        expect(page).to have_selector :region, "Section region"
+        expect(page).to have_selector :region, "Section region", count: 1
       end
 
       it "does not find differing on [aria-labelledby]" do
@@ -45,6 +49,23 @@ describe "region selector" do
         HTML
 
         expect(page).to have_no_selector :region, "Not the right locator"
+      end
+
+      it "does not find with empty aria-label" do
+        render <<~HTML
+          <section aria-label=" ">Content</section>
+        HTML
+
+        expect(page).to have_no_selector :region
+      end
+
+      it "does not find with empty aria-labelledby" do
+        render <<~HTML
+          <section aria-labelledby="id">Content</section>
+          <div id="id"></div>
+        HTML
+
+        expect(page).to have_no_selector :region
       end
     end